Extract sorting question bank sorting functionality to a general utility function
