+++ server/app/assets/javascripts/questionBank.ts
@@ -1,5 +1,6 @@
 /** The question bank controller is responsible for manipulating the question bank. */
-import {assertNotNull} from './util'
+import { assertNotNull } from './util'
+import { sortSelectorElements } from './sort_selector'
 
 class QuestionBankController {
   static readonly FILTER_ID = 'question-bank-filter'
@@ -106,7 +107,7 @@ class QuestionBankController {
         container.classList.add('hidden')
         QuestionBankController.makeBodyScrollable()
       },
-      {once: true},
+      { once: true },
     )
     container.classList.add(QuestionBankController.QUESTION_BANK_HIDDEN)
     const url = new URL(location.href)
@@ -147,7 +148,7 @@ class QuestionBankController {
       questionElement.classList.toggle(
         'hidden',
         filterString.length > 0 &&
-          !questionFilterText.toUpperCase().includes(filterString),
+        !questionFilterText.toUpperCase().includes(filterString),
       )
     })
   }
@@ -156,50 +157,7 @@ class QuestionBankController {
    * Sort questions in the question bank based on the criteria selected from the dropdown.
    */
   private static sortQuestions() {
-    const questionBankSort = document.getElementById(
-      QuestionBankController.SORT_SELECT_ID,
-    ) as HTMLSelectElement
-
-    const questionSublists = document.querySelectorAll('.cf-sortable-questions')
-    if (!questionBankSort || !questionSublists) {
-      return
-    }
-
-    questionSublists.forEach((questionSublist) => {
-      const questions: HTMLElement[] = Array.from(
-        questionSublist.querySelectorAll('.cf-question-bank-element'),
-      )
-
-      const sortedQuestions = questions.sort((elementA, elementB) => {
-        // questionBankSort.value is expected to be of the format "<data_attribute_name>-<asc|desc>".
-        // Attribute names and order suffix are defined in QuestionSortOption.java.
-        const [attrName, order] = questionBankSort.value.split('-')
-        // Get the data attribute whose name matches the selected sort option so that it can be used to compare the elements.
-        const attrA: string | null = elementA.getAttribute('data-' + attrName)
-        const attrB: string | null = elementB.getAttribute('data-' + attrName)
-        if (!attrA || !attrB) {
-          return 0
-        }
-
-        const compare = function (a: string, b: string): number {
-          switch (attrName) {
-            case 'lastmodified': {
-              const dateA = new Date(a)
-              const dateB = new Date(b)
-              return dateA.getTime() - dateB.getTime()
-            }
-            default:
-              // Default sort is a string sort.
-              return a.localeCompare(b)
-          }
-        }
-        return order == 'asc' ? compare(attrA, attrB) : compare(attrB, attrA)
-      })
-
-      sortedQuestions.forEach((q) => {
-        questionSublist.appendChild(q)
-      })
-    })
+    sortSelectorElements(QuestionBankController.SORT_SELECT_ID, "." + QuestionBankController.QUESTION_CLASS, '.cf-sortable-questions');
   }
 }
